- Made it so we can now select items in the quest. - Also, brought the framework under the test harness.

+++ trunk/Aether RPG/src/com/aether/present/game/InGameWorldWindow.java
@@ -333,6 +333,9 @@ public class InGameWorldWindow extends BasicGameState {
 	@Override
 	public void cleanup() {
 		super.cleanup();
+		if (textureState == null) {
+			return;
+		}
 		textureState.clearTextures();
 		textureState.deleteAll();
 		textureState = null;
